web/pottermania_part2/challenge_code/app/requirements.txt
blinker==1.9.0
cachelib==0.9.0
click==8.1.8
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
PyJWT==2.10.1
redis==5.2.1
Werkzeug==3.1.3
